Short Summary.  Why does it even have this line?
I did some work on the guard searching.  After the last few hours, all I know is that I'm having problems with the raycasting.  It may have something to do with the distance formulas taking so long to calculate.  Those seem to be wrong after a period of time.  In fact, when the two are right next to each other, the distance just grows.  Not sure what that is about, but that causes errors.   Even before that though, the player is definitely in view, but the ray cast doesn't always seem to hit.  Need to figure out what's going on.  Could be a problem with how big the items are, but I don't know.  Height seems to matter, but it's sometimes seeing the player, and sometimes not.  And I can't determine when or why.  Also, it seems like the states might be changing to quickly for this to be working.  Again, could have to deal with how long it takes this thing to calculate all the stuff.  But it doesn't seem like that should really be the issue.