prism tooltips and grabbable objects
prisms all have tooltips and grabbable objects now. There are some glitches, tooltips are defaulting to on and then they work as expected after the ontriggerexit is tripped the first time. I also need to make it so that you can't grab the prisms which are supposed to be stationary (eg locked in place)