Fixes
I fixed the issue in which players starting the game couldn't see the rooms already created.  I also fixed the issue in which players would be automatically sent to the end of the game screen.  I noticed that the city wasn't showing up in this build though.  Not sure what that's about.  Also, it seems like there was something else I fixed, but I can't quite recall what it might have been.

Once, I ran into the issue where none of the networking peices were built.  I have no idea what caused that, and it only happened to me once.  I feel like there should be a way to look into that and ensure it happens, but I don't know what that way would be.  Maybe someone else in class has dealt with this before?